+++ R/hyper_tbl_cube.R
@@ -25,7 +25,7 @@
 #' @examples
 #' f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
 #' l3file <- system.file("extdata/oceandata", f, package= "tidync")
-#' (cube <- hyper_tbl_cube(tidync(l3file) %>%
+#' (cube <- hyper_tbl_cube(tidync(l3file) |>
 #' activate(chlor_a), lon = lon > 107, lat = abs(lat) < 30))
 #' ufile <- system.file("extdata", "unidata", "test_hgroups.nc", 
 #'  package = "tidync", mustWork = TRUE)
@@ -51,14 +51,14 @@ hyper_tbl_cube <- function(x, ..., force = FALSE) {
 hyper_tbl_cube.tidync <- function(x, ..., force = FALSE) {
   active_names <- tibble::tibble(dim = as.integer(
     gsub("^D", "", unlist(strsplit(active(x), ",")))))
-  dim_names <- active_names %>% 
-    inner_join(x[["dimension"]] %>% 
-                 dplyr::filter(active), c("dim" = "id")) %>% 
+  dim_names <- active_names |> 
+    inner_join(x[["dimension"]] |> 
+                 dplyr::filter(active), c("dim" = "id")) |> 
     dplyr::pull(.data$name)
   trans <- x[["transforms"]][dim_names]
   lfun <- function(inm) {
-    trans[[inm]] %>% 
-      dplyr::filter(.data$selected) %>% 
+    trans[[inm]] |> 
+      dplyr::filter(.data$selected) |> 
       dplyr::pull(inm)
   }
   ldims <- lapply(dim_names, lfun)
@@ -69,7 +69,7 @@ hyper_tbl_cube.tidync <- function(x, ..., force = FALSE) {
 #' @name hyper_tbl_cube
 #' @export
 hyper_tbl_cube.character <- function(x, ..., force = FALSE) {
-  tidync(x) %>% hyper_filter(...) %>% hyper_tbl_cube(force = force)
+  tidync(x) |> hyper_filter(...) |> hyper_tbl_cube(force = force)
 }