+++ R/netcdf-utils.R
@@ -12,7 +12,6 @@
 #'
 #' @importFrom ncdf4 nc_open nc_close ncvar_get
 #' @importFrom RNetCDF open.nc close.nc var.get.nc
-#' @importFrom purrr safely
 #' @export
 nc_get <- function(x, v, test = FALSE) {
   UseMethod("nc_get")
@@ -22,10 +21,9 @@ nc_get.character <- function(x, v, test = FALSE) {
   if (!test) {
     con <- RNetCDF::open.nc(x)
     on.exit(RNetCDF::close.nc(con), add = TRUE)
-  
-    safe_get <- purrr::safely(nc_get.NetCDF)
-    val <- safe_get(con, v)
-    if (!is.null(val$result)) return(val$result)
+
+    val <- tryCatch(nc_get.NetCDF(con, v), error = function(e) NULL)
+    if (!is.null(val)) return(val)
   } else {
   con4 <- NULL
   ## issue #119
@@ -33,12 +31,11 @@ nc_get.character <- function(x, v, test = FALSE) {
   con4 <- ncdf4::nc_open(x, readunlim = FALSE, verbose = FALSE, 
                          auto_GMT = FALSE, suppress_dimvals = TRUE))
   on.exit(ncdf4::nc_close(con4), add = TRUE)
-  safe_get4 <- purrr::safely(nc_get.ncdf4)
-  val <- safe_get4(con4, v)
-  if (is.null(val[["result"]])) {
+  val <- tryCatch(nc_get.ncdf4(con4, v), error = function(e) NULL)
+  if (is.null(val)) {
     stop(sprintf("no variable found %s", v))
   } else {
-    return(val[["result"]])
+    return(val)
   }
   
   }