+++ tests/testthat/test-files-sources.R
@@ -42,20 +42,20 @@ test_that("verbs have a character method", {
   f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
   l3file <- system.file("extdata/oceandata", f, 
                         package= "tidync", mustWork = TRUE)
-  hyper_tibble(l3file, lon = lon < -140, lat = lat > 85) %>% 
+  hyper_tibble(l3file, lon = lon < -140, lat = lat > 85) |> 
     expect_s3_class("tbl_df")
-  hyper_filter(l3file, lon = lon < -140, lat = lat > 85) %>% 
+  hyper_filter(l3file, lon = lon < -140, lat = lat > 85) |> 
     expect_s3_class("tidync")
   expect_warning(hyper_slice(l3file, 
                              lon = lon < -140, 
                              lat = lat > 85, 
-                             select_var = "chlor_a")) %>% 
+                             select_var = "chlor_a")) |> 
     expect_type("list")
   expect_silent(hyper_array(l3file, lon = lon < -140, lat = lat > 85, 
-                            select_var = "chlor_a")) %>% 
+                            select_var = "chlor_a")) |> 
     expect_type("list")
   
-  hyper_tbl_cube(l3file, lon = lon < -140, lat = lat > 85) %>% 
+  hyper_tbl_cube(l3file, lon = lon < -140, lat = lat > 85) |> 
     expect_s3_class("tbl_cube")
   
 })
@@ -63,10 +63,10 @@ test_that("RNetCDF fall back works", {
   skip_on_cran()
   ufile <- system.file("extdata", "unidata", "madis-hydro.nc", 
                        package = "tidync", mustWork = TRUE)
-  nc_get(ufile, "invTime") %>% expect_is("array") %>% 
-    expect_length(1176) #%>% expect_type("double")
-  nc_get(ufile, "invTime", test = TRUE) %>% expect_is("array") %>% 
-    expect_length(1176) #%>% expect_type("integer")
+  nc_get(ufile, "invTime") |> expect_is("array") |> 
+    expect_length(1176) #|> expect_type("double")
+  nc_get(ufile, "invTime", test = TRUE) |> expect_is("array") |> 
+    expect_length(1176) #|> expect_type("integer")
   
   expect_output({
   expect_error(nc_get(ufile, "notavariable"))